web/news_board/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from .views import PostViewSet, CommentViewSet, Upvotes

router = DefaultRouter()
router.register(r"posts", PostViewSet)
router.register(r"comments", CommentViewSet)

urlpatterns = [
    path("upvotes/<int:pk>/", Upvotes.as_view()),
    path("", include(router.urls)),
]
