web/requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.3.0
black==20.8b1
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.3
djangorestframework==3.12.2
flake8==3.8.4
importlib-metadata==2.0.0
kombu==5.0.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
redis==3.5.3
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
